requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
httpx==0.27.0

sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

pydantic==2.7.1
python-dotenv==1.0.1